+++ cocos/2d/CCFontFreeType.cpp
@@ -99,7 +99,7 @@ bool FontFreeType::createFontObject(const std::string &fontName, int fontSize)
 {
     FT_Face face;
 
-    long len = 0;
+    ssize_t len = 0;
     _ttfData = FileUtils::getInstance()->getFileData(fontName.c_str(), "rb", &len);
     if (!_ttfData)
         return false;
